requirements.txt
langchain==0.1.0
langchain-community==0.0.10
qdrant-client==1.7.0
gradio==4.7.1
sentence-transformers>=2.3.0
python-dotenv==1.0.0
numpy==1.24.3

